requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
google-cloud-vision>=3.7.4
google-api-core>=2.19.0
requests>=2.32.3
openai>=1.47.0
python-dotenv>=1.0.1
pydantic>=2.9.2
aiofiles>=23.2.1
python-multipart>=0.0.9
